+++ CHANGELOG.md
@@ -52,6 +52,7 @@ You can find its changes [documented below](#070---2021-01-01).
 - x11: Implement primary_clipboard ([#1867] by [@psychon])
 - x11: Set WM_CLASS property ([#1868] by [@psychon])
 - Expose `RawWindowHandle` for `WindowHandle` under the `raw-win-handle` feature ([#1828] by [@djeedai])
+- Widget/Slider: Add stepping functionality ([#1875] by [@raymanfx])
 
 ### Changed